// src/pages/admin/types.ts
export interface AdminTaskTypeItem {
  id: string;
  type1: string;
  type2: string;
  displayLabel: string;
  displayOrder: number;
  requiresServiceGroup: boolean;
  isActive: boolean;
}

export interface AdminTaskTypePayload {
  id?: string;
  type1: string;
  type2: string;
  displayLabel: string;
  displayOrder: number;
  requiresServiceGroup: boolean;
  isActive: boolean;
}